added CustomUserDetailsService

+++ src/main/java/com/cts/ClinicManagement/entity/Roles.java
@@ -3,7 +3,6 @@ package com.cts.ClinicManagement.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -16,8 +15,8 @@ import lombok.Setter;
 public class Roles {
 
 	@Id
-	private int id;
+	private Long id;
 	private String role;
 	
-	
+
 }
